# packages/contracts/openapi/brandcanvas.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/seller/products": {
      "get": {
        "operationId": "Catalog_findMany",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "active",
                "inactive",
                "archived"
              ]
            }
          },
          {
            "name": "stockStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "in_stock",
                "low_stock",
                "out_of_stock"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "summary": "List products and current inventory for the authenticated seller store",
        "tags": [
          "Seller Catalog"
        ]
      },
      "post": {
        "operationId": "Catalog_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "summary": "Create a product, default variant, and inventory item",
        "tags": [
          "Seller Catalog"
        ]
      }
    },
...
    "/api/v1/seller/inventory": {
      "get": {
        "operationId": "InventoryManagement_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stockStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "in_stock",
                "low_stock",
                "out_of_stock"
              ]
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Inventory"
        ]
      }
    },
    "/api/v1/seller/inventory/products/{productId}": {
      "get": {
        "operationId": "InventoryManagement_getProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductInventoryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Inventory"
        ]
      }
    },
    "/api/v1/seller/inventory/{inventoryItemId}/movements": {
      "get": {
        "operationId": "InventoryManagement_listMovements",
        "parameters": [
          {
            "name": "inventoryItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "movementType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "initial_stock",
                "purchase",
                "return",
                "order_cancelled",
                "damaged",
                "manual_increase",
                "manual_decrease",
                "set_quantity",
                "reservation",
                "reservation_release",
                "reservation_expiry",
                "sale",
                "cancellation_restore",
                "return_restore",
                "correction"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryMovementListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Inventory"
        ]
      }
    },
    "/api/v1/seller/inventory/{inventoryItemId}/adjustments": {
      "post": {
        "operationId": "InventoryManagement_adjust",
        "parameters": [
          {
            "name": "inventoryItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryAdjustmentRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Inventory"
        ]
      }
    },
    "/api/v1/seller/inventory/{inventoryItemId}/threshold": {
      "patch": {
        "operationId": "InventoryManagement_updateThreshold",
        "parameters": [
          {
            "name": "inventoryItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLowStockThresholdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Inventory"
        ]
      }
    },
    "/api/v1/seller/inventory/{inventoryItemId}/reservations": {
      "post": {
        "operationId": "InventoryManagement_reserve",
        "parameters": [
          {
            "name": "inventoryItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveInventoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryReservationOperationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Inventory"
        ]
      }
    },
    "/api/v1/seller/inventory/reservations/{reservationId}/release": {
      "post": {
        "operationId": "InventoryManagement_release",
        "parameters": [
          {
            "name": "reservationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryReservationOperationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Inventory"
        ]
      }
    },
    "/api/v1/seller/inventory/reservations/{reservationId}/expire": {
      "post": {
        "operationId": "InventoryManagement_expire",
        "parameters": [
          {
            "name": "reservationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryReservationOperationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Inventory"
        ]
      }
    },
    "/api/v1/seller/inventory/reservations/{reservationId}/cancel": {
      "post": {
        "operationId": "InventoryManagement_cancel",
        "parameters": [
          {
            "name": "reservationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryReservationOperationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Inventory"
        ]
      }
    },
    "/api/v1/seller/inventory/reservations/{reservationId}/convert": {
      "post": {
        "operationId": "InventoryManagement_convert",
        "parameters": [
          {
            "name": "reservationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryReservationOperationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "brandcanvas_access": []
          }
        ],
        "tags": [
          "Seller Inventory"
        ]
      }
    },
...
  },
  "info": {
    "title": "BrandCanvas API",
    "description": "Central API contract for the BrandCanvas multi-tenant commerce platform.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "brandcanvas_access"
      }
    },
    "schemas": {
...
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "inventoryItemId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "inactive",
              "archived"
            ]
          },
          "sku": {
            "type": "string"
          },
          "priceMinor": {
            "type": "number",
            "example": 250000
          },
          "compareAtPriceMinor": {
            "type": "number",
            "nullable": true,
            "example": 300000
          },
          "stockQuantity": {
            "type": "number"
          },
          "reservedQuantity": {
            "type": "number"
          },
          "availableQuantity": {
            "type": "number"
          },
          "lowStockThreshold": {
            "type": "number"
          },
          "stockStatus": {
            "type": "string",
            "enum": [
              "in_stock",
              "low_stock",
              "out_of_stock"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "variantId",
          "inventoryItemId",
          "name",
          "slug",
          "status",
          "sku",
          "priceMinor",
          "stockQuantity",
          "reservedQuantity",
          "availableQuantity",
          "lowStockThreshold",
          "stockStatus",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponseDto"
            }
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Classic T-Shirt"
          },
          "description": {
            "type": "string",
            "example": "Premium cotton t-shirt."
          },
          "sku": {
            "type": "string",
            "example": "TSHIRT-BLK-M"
          },
          "priceMinor": {
            "type": "number",
            "example": 250000,
            "minimum": 0
          },
          "compareAtPriceMinor": {
            "type": "number",
            "example": 300000,
            "minimum": 0,
            "nullable": true
          },
          "initialStock": {
            "type": "number",
            "default": 0,
            "minimum": 0
          },
          "lowStockThreshold": {
            "type": "number",
            "default": 5,
            "minimum": 0,
            "maximum": 1000000
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "inactive",
              "archived"
            ],
            "default": "draft"
          }
        },
        "required": [
          "name",
          "sku",
          "priceMinor",
          "initialStock",
          "lowStockThreshold",
          "status"
        ]
      },
...
      "InventoryItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "productName": {
            "type": "string"
          },
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "variantTitle": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "stockQuantity": {
            "type": "number"
          },
          "reservedQuantity": {
            "type": "number"
          },
          "availableQuantity": {
            "type": "number"
          },
          "lowStockThreshold": {
            "type": "number"
          },
          "stockStatus": {
            "type": "string",
            "enum": [
              "in_stock",
              "low_stock",
              "out_of_stock"
            ]
          },
          "isAvailable": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "productId",
          "productName",
          "variantId",
          "variantTitle",
          "sku",
          "stockQuantity",
          "reservedQuantity",
          "availableQuantity",
          "lowStockThreshold",
          "stockStatus",
          "isAvailable",
          "createdAt",
          "updatedAt"
        ]
      },
      "InventoryListResponseDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "default": 1,
            "minimum": 1
          },
          "pageSize": {
            "type": "number",
            "default": 20,
            "minimum": 1,
            "maximum": 100
          },
          "search": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryItemResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "totalPages"
        ]
      },
      "ProductInventoryResponseDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryItemResponseDto"
            }
          }
        },
        "required": [
          "productId",
          "items"
        ]
      },
      "InventoryMovementResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "movementType": {
            "type": "string",
            "enum": [
              "initial_stock",
              "purchase",
              "return",
              "order_cancelled",
              "damaged",
              "manual_increase",
              "manual_decrease",
              "set_quantity",
              "reservation",
              "reservation_release",
              "reservation_expiry",
              "sale",
              "cancellation_restore",
              "return_restore",
              "correction"
            ]
          },
          "quantityDelta": {
            "type": "number"
          },
          "stockBefore": {
            "type": "number"
          },
          "stockAfter": {
            "type": "number"
          },
          "reservedBefore": {
            "type": "number"
          },
          "reservedAfter": {
            "type": "number"
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "referenceType": {
            "type": "string",
            "nullable": true
          },
          "referenceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "actorName": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "movementType",
          "quantityDelta",
          "stockBefore",
          "stockAfter",
          "reservedBefore",
          "reservedAfter",
          "metadata",
          "createdAt"
        ]
      },
      "InventoryMovementListResponseDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "default": 1,
            "minimum": 1
          },
          "pageSize": {
            "type": "number",
            "default": 20,
            "minimum": 1,
            "maximum": 100
          },
          "search": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryMovementResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "totalPages"
        ]
      },
      "InventoryAdjustmentRequestDto": {
        "type": "object",
        "properties": {
          "operation": {
            "type": "string",
            "enum": [
              "increase",
              "decrease",
              "set"
            ]
          },
          "quantity": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000
          },
          "reason": {
            "type": "string",
            "minLength": 3,
            "maxLength": 500
          },
          "idempotencyKey": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          }
        },
        "required": [
          "operation",
          "quantity",
          "reason"
        ]
      },
      "UpdateLowStockThresholdDto": {
        "type": "object",
        "properties": {
          "lowStockThreshold": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000
          },
          "reason": {
            "type": "string",
            "minLength": 3,
            "maxLength": 500
          },
          "idempotencyKey": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          }
        },
        "required": [
          "lowStockThreshold",
          "reason"
        ]
      },
      "ReserveInventoryDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 1000000
          },
          "referenceType": {
            "type": "string",
            "maxLength": 50
          },
          "referenceId": {
            "type": "string",
            "format": "uuid"
          },
          "idempotencyKey": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "quantity",
          "referenceType",
          "referenceId",
          "idempotencyKey",
          "expiresAt"
        ]
      },
      "InventoryReservationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "inventoryItemId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "converted",
              "released",
              "expired",
              "cancelled"
            ]
          },
          "referenceType": {
            "type": "string"
          },
          "referenceId": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "releasedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "inventoryItemId",
          "quantity",
          "status",
          "referenceType",
          "referenceId",
          "expiresAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "InventoryReservationOperationResponseDto": {
        "type": "object",
        "properties": {
          "inventory": {
            "$ref": "#/components/schemas/InventoryItemResponseDto"
          },
          "reservation": {
            "$ref": "#/components/schemas/InventoryReservationResponseDto"
          }
        },
        "required": [
          "inventory",
          "reservation"
        ]
      },
...
    }
  }
}